requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.0.4
idna==2.10
isort==5.8.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.3
optional-django==0.3.0
postgres==3.0.0
psycopg2==2.7.7
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
react==4.3.0
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
